add new files from me

+++ src/main/java/com/mohamed/models/Constants.java
@@ -2,6 +2,6 @@ package com.mohamed.models;
 
 public class Constants {
 
-    public static final int MAX_BOOKS_ISSUED_TO_A_USER = 5;
-    public static final int MAX_LENDING_DAYS = 10;
+    public static final long MAX_BOOKS_ISSUED_TO_A_USER = 5;
+    public static final long MAX_LENDING_DAYS = 10;
 }
